requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mongoengine==0.27.0
pymongo==4.6.0
PyJWT==2.8.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
bcrypt==4.1.2
